requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==11.3.0
filelock==3.4.2
flake8==4.0.1
mccabe==0.6.1
mongoengine==0.23.1
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==3.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
tzdata==2021.5
